Move PerfHud into shared chat module
