+++ app-chat/src/main/java/com/example/kokoro/chat/LlmInference.kt
@@ -5,7 +5,6 @@ import com.google.mediapipe.tasks.genai.llminference.LlmInference
 import com.google.mediapipe.tasks.genai.llminference.LlmInference.LlmInferenceOptions
 import java.io.File
 import com.example.kokoro82m.utils.DebugLogger
-import com.example.kokoro.galleryport.PerfHud
 
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.Dispatchers
